Dispatch temporary tablespace id to all Gangs.
If a query has multiple Gangs, the secondary Gang cannot get correct tmporary tablespace ID.

Consider this SQL:

```
    create table tn_a(id int) distributed by (id); insert into tn_a values (1), (2);
    create temp table tn_a_tmp(a int) distributed replicated; insert into tn_a_tmp values(1);

    create or replace function fun(sql text, a oid) returns bigint AS '
        return plpy.execute(sql).nrows() + a
    ' language plpython3u stable;

    create table tn_a_new as with c as (
        select fun('select * from tn_a_tmp', s.id) from tn_a s
    ) select 1 from c;
```

The key is to reader gang to access temp table via text-sql execute, this can only happen when
UDF executing text SQL and the temp table is replicated.

In this scenario happened. the frist `create temp table` will create a
temporary schema then dispatch to primary QE. both QD and QE has a
correct myTempNamespaceID. when executing the 2-Gang query. the
secondary Gang will boot dynamic with myTempNamespaceID = InvalidOid.

In the end, the secondary Gang scaning on temporary table. will throw a
'table not exist' error with out checking temporary schema.

The method to solve this problem is the same as numsegments in GPDB.
We re-dispatch the temporary namespace OID in each query. By adding 8byte
to be sent in the MPP query message. The temporary namespace OID cannot be
changed after it has been set, same as PG.

Signed-off-by: Sasasu <i@sasa.su>

use buildGpQueryString to dispatch temp NS id